refactor: implement generic API pattern and improve type annotations
- Add generic ApiBase class to eliminate code duplication between sync/async implementations
- Refactor all API modules (issues, comments, fields, search) to use unified generic pattern
- Add comprehensive type annotations to all example files
- Update Makefile to include examples directory in type checking
- Remove redundant API module structure in favor of unified approach
