--- src/jira2py/api/issue_comments/__init__.py
@@ -1,4 +1,0 @@
-from .issue_comments_async import IssueCommentsAsync
-from .issue_comments_sync import IssueComments
-
-__all__ = ["IssueCommentsAsync", "IssueComments"]
