--- src/jira2py/api/issue_fields/__init__.py
@@ -1,4 +1,0 @@
-from .issue_fields_async import IssueFieldsAsync
-from .issue_fields_sync import IssueFields
-
-__all__ = ["IssueFieldsAsync", "IssueFields"]
