--- src/jira2py/api/issue_search/__init__.py
@@ -1,4 +1,0 @@
-from .issue_search_async import IssueSearchAsync
-from .issue_search_sync import IssueSearch
-
-__all__ = ["IssueSearchAsync", "IssueSearch"]
