--- src/jira2py/api/issues/__init__.py
@@ -1,4 +1,0 @@
-from .issues_async import IssuesAsync
-from .issues_sync import Issues
-
-__all__ = ["Issues", "IssuesAsync"]
