+++ src/jira2py/api/jira_api_async.py
@@ -7,7 +7,7 @@ from .issue_fields import IssueFieldsAsync
 from .issue_search import IssueSearchAsync
 from .issues import IssuesAsync
 from .jira_api_base import JiraAPIBase
-from jira2py.client import JiraClientAsync, JiraCredentials
+from jira2py.client import JiraClientAsync
 
 
 class JiraAPIAsync(JiraAPIBase):
@@ -24,25 +24,9 @@ class JiraAPIAsync(JiraAPIBase):
         ...     fields = await api.fields.get_fields()
     """
 
-    def __init__(
-        self,
-        credentials: JiraCredentials | None = None,
-        url: str | None = None,
-        username: str | None = None,
-        api_token: str | None = None,
-    ) -> None:
-        """Initialize the asynchronous Jira API facade.
-
-        Args:
-            credentials: JIRA authentication credentials. If None, credentials will be
-                created from the provided parameters or environment variables.
-            url: Base URL of the JIRA instance. Only used if credentials is None.
-            username: JIRA username. Only used if credentials is None.
-            api_token: JIRA API token. Only used if credentials is None.
-        """
-        super().__init__(url, username, api_token)
-        # Single shared client for all modules
-        self._client = JiraClientAsync(self.credentials)
+    def _create_client(self) -> JiraClientAsync:
+        """Create the async client instance."""
+        return JiraClientAsync(self.credentials)
 
     @property
     def issues(self) -> IssuesAsync:
