+++ src/jira2py/client/__init__.py
@@ -3,5 +3,6 @@
 from .client_async import JiraClientAsync
 from .credentials import JiraCredentials
 from .client_sync import JiraClientSync
+from .factory import JiraClientFactory
 
-__all__ = ["JiraCredentials", "JiraClientSync", "JiraClientAsync"]
+__all__ = ["JiraCredentials", "JiraClientSync", "JiraClientAsync", "JiraClientFactory"]
